requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
flask-cors
Flask==1.1.2
lxml
html5lib
Flask-Compress==1.8.0
Flask-SeaSurf==0.2.2
gunicorn==20.0.4
Jinja2==2.11.2
PyYAML==5.3.1
requests==2.25.0
urllib3==1.26.2
Werkzeug==1.0.1
Flask-Login==0.4.1
Flask-Session==0.3.1
Flask_SocketIO
itsdangerous==1.1.0
MarkupSafe==1.1.0
python-engineio
python-socketio
six==1.11.0
